Add policies to fields and make association consistent
Policies allow conditions to be added to fields that determine whether or not
they may be seen. Only one policy is provided out of the box: it checks if a
key on the logged in user is the same as a key on the current data.

Finally, the associations now also have a reference to the query builder they
were declared in, just like the fields. If the policies are ever added to
associations as well, this will allow policies access to the entire assoc flow.