Allow the query builder to accept manual render specifications
A specification can now be passed to the query builder that accepts input
similar to that of a request which will then be used for fetching / building /
rendering. This allows the querybuilder to be used as a general data fetcher and
renderer without having to build Request objects all the time.