WIP: Implement a documentated validation through builder
Allow the validation to the endpoints to be declared through a validation rule
builder that can then also be used to generate documentation from these defined
rules.

Each rule is stored as an object that has a documentation line associated with
it. This method, while repetitive, ensures that the validation rules remain
compatible with Laravel's validator while still remaining discoverable through
code.