Add automatic controller and routing for a schema (#25)
* WIP: Add automatic controller and routing for a schema

Using a macro on Laravel's Router, we can quickly define routes for a specific
query builder. Using this approach, we can easily re-use most of the
functionality from Laravel (e.g. by wrapping the route in a group with some
middleware).

The routes can then be send to a generic controller that knows about the query
builder that we've defined in the route. It uses a somewhat sneaky method for
passing this metadata to the controller; the extra metadata is passed in the
array that Laravel uses to store the action for that route.

The next step would be to add relationships to the route and the generic
controller, such that routes like:

/users/1/relationships/posts

would function out of the box.

* Add tests for the api routes being constructed

* Add composer analyse to travis

* Add more tests and split route macro to separate provider

* Rewrite routing with new scope callback

* Add show/index to the generic controller

* Rename query builder to schema